+++ CHANGES.md
@@ -9,6 +9,8 @@
 -- `Get-ChildItem` does not fail in the empty directory but still fails in not empty
 - [v5-LiteralPath-Recurse-ignores-Include](Cmdlets\Get-ChildItem\v5-LiteralPath-Recurse-ignores-Include)
 -- Fixed the v5 regression
+- [PowerShell.exe\Unexpected-start-location](PowerShell.exe\Unexpected-start-location)
+-- Fixed
 
 ### v6-RC